services/document-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
python-multipart==0.0.6

# Storage
boto3==1.34.0
google-cloud-storage==2.13.0

# OCR
google-cloud-vision==3.4.5
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0

# Parsing
python-dateutil==2.8.2
